fix(test): test for serializers are separated
