+++ tests/data/serialize/tsadra/P000801/P000801.opf/base/v001.txt
@@ -5,6 +5,9 @@
 ཐུགས་དང་མཛད་པ་འཕྲིན་ལས་ལྷུན་
 གྲུབ་བསམ་གྱིས་མི་ཁྱབ་བློ་ལས་རྣམ་འ
 དས་འགྲོ་བའི་དཔལ་གྱུར་ཆོས་ཀྱི་རྗེ་ལ་འདུད། །
+ཐུགས་དང་མཛད་པ་འཕྲིན་ལས་ལྷུན་
+གྲུབ་བསམ་གྱིས་མི་ཁྱབ་བློ་ལས་རྣམ་འ
+དས་འགྲོ་བའི་དཔལ་གྱུར་ཆོས་ཀྱི་རྗེ་ལ་འདུད། །
 དུས་གསུམ་འགར་མེད་འོད་གསལ་སྒྲིབ་གསུམ་རབ་སྤངས་རྣམ་དག་ཐུགས་ཀྱི་དཀྱིལ་འཁོར་བསམ་གྱིས་མི་ཁྱབ་ཡེ་ཤེས་དུས་རྣམས་ཀུན་ལ་ཆགས་ཐོགས་མེད་པར་འཇུག །ཡོན་ཏན་རྒྱ་མཚོ་སྐུ་གསུང་ཐུགས་དང་མཛད་པ་འཕྲིན་ལས་ལྷུན་གྲུབ་བསམ་གྱིས་མི་ཁྱབ་བློ་ལས་རྣམ་འདས་འགྲོ་བའི་དཔལ་གྱུར་ཆོས་ཀྱི་རྗེ་ལ་འདུད། །
