+++ tests/data/serialize/tsadra/P000801/P000801.opf/layers/v001/Sabche.yml
@@ -4,5 +4,5 @@ revision: '00014'
 annotations:
   3f56858201fc44e8808841286131a475:
     span:
-      start: 466
-      end: 501
+      start: 572
+      end: 607
