+++ tests/data/serialize/tsadra/editor_serializer_output.html
@@ -3,5 +3,5 @@
 <title></title>
 </head>
 <body>
-<p><span class="book-title" id="352bdde9266843c8a2ef739617c8e25a">༄། །འབྲི་གུང་ཚོགས་ཆོས་ཆེན་མོ།</span></p><p><span class="sub-title" id="57fb594c15f749949a486a690367908e">༼ཚགས་ཆོས་ཀྱི་རིམ་པ་ལེགས་པར་བཤད་པ་ཡིད་བཞིན་ནོར་བུ་ཞེས་བྱ་བ་བཞུགས་སོ།།༽</span></p><p><span class="author" id="0688c562cb8b48438987d82dc05f3dfb">ཚོགས་གཉིས་</span></p><p><span class="citation" id="e54f06a0254e4e1f9f72552c514841b3">མཐར་ཕྱིན་མཚན་དང་དཔེ་བྱད་རིན་ཆེན་རི་བོ་ལྟར་</span>མཛེས་སྐུ་ཡི་དཀྱིལ་འཁོར་བསམ་གྱིས་མི་<span class="citation" id="a3cacfe626734376b5a45ade275de779">ཁྱབ་མཁའ་མཉམ་སེམས་ཅན་ཁམས་དང་མཉམ་པར་འཇུག །</span>ཆོས་རྣམས་གདོད་ནས་མ་སྐྱེས་སྟོང་པར་སྟོན་པའི་གསུང་གི་དཀྱིལ་འཁོར་བསམ་གྱིས་མི་ཁྱབ་ཐེ་ཚོམ་གཅོད་ཅིང་སེམས་ཅན་<span class="yigchung" id="4cdb2351350a498f8c07a84c9781835e">རྣམས་ཀྱི་སྐད་དང་མཐུན་པར་འཇུག །</span></p><p><span class="root-text" id="e79dfe48ae3940078d202c9f592ac864">ཐུགས་དང་མཛད་པ་འཕྲིན་ལས་ལྷུན་</span></p><p><span class="root-text" id="e79dfe48ae3940078d202c9f592ac8641">གྲུབ་བསམ་གྱིས་མི་ཁྱབ་བློ་ལས་རྣམ་འ</span></p><p><span class="root-text" id="e79dfe48ae3940078d202c9f592ac8642">དས་འགྲོ་བའི་དཔལ་གྱུར་ཆོས་ཀྱི་རྗེ་ལ་འདུད། །</span></p><p><span class="sabche" id="3f56858201fc44e8808841286131a475">དུས་གསུམ་འགར་མེད་འོད་གསལ་སྒྲིབ་གསུམ་</span>རབ་སྤངས་རྣམ་དག་ཐུགས་ཀྱི་དཀྱིལ་འཁོར་བསམ་གྱིས་མི་ཁྱབ་ཡེ་ཤེས་དུས་རྣམས་ཀུ<span class="root-text" id="de30c1c91ce44217a6e0cacdcab6a404">ན་ལ་ཆགས་ཐོགས་མེད་པར་འཇུག །</span>ཡོན་ཏན་རྒྱ་མཚོ་སྐུ་གསུང་ཐུགས་དང་མཛད་པ་འཕྲིན་ལས་ལྷུན་གྲུབ་བསམ་གྱིས་མི་ཁྱབ་བློ་ལས་རྣམ་འདས་འགྲོ་བའི་དཔལ་གྱུར་ཆོས་ཀྱི་རྗེ་ལ་འདུད། །</p><p></p><p></p><p></p></body>
+<p><span class="book-title" id="352bdde9266843c8a2ef739617c8e25a">༄། །འབྲི་གུང་ཚོགས་ཆོས་ཆེན་མོ།</span></p><p><span class="sub-title" id="57fb594c15f749949a486a690367908e">༼ཚགས་ཆོས་ཀྱི་རིམ་པ་ལེགས་པར་བཤད་པ་ཡིད་བཞིན་ནོར་བུ་ཞེས་བྱ་བ་བཞུགས་སོ།།༽</span></p><p><span class="author" id="0688c562cb8b48438987d82dc05f3dfb">ཚོགས་གཉིས་</span></p><p><span class="citation" id="e54f06a0254e4e1f9f72552c514841b3">མཐར་ཕྱིན་མཚན་དང་དཔེ་བྱད་རིན་ཆེན་རི་བོ་ལྟར་</span>མཛེས་སྐུ་ཡི་དཀྱིལ་འཁོར་བསམ་གྱིས་མི་<span class="citation" id="a3cacfe626734376b5a45ade275de779">ཁྱབ་མཁའ་མཉམ་སེམས་ཅན་ཁམས་དང་མཉམ་པར་འཇུག །</span>ཆོས་རྣམས་གདོད་ནས་མ་སྐྱེས་སྟོང་པར་སྟོན་པའི་གསུང་གི་དཀྱིལ་འཁོར་བསམ་གྱིས་མི་ཁྱབ་ཐེ་ཚོམ་གཅོད་ཅིང་སེམས་ཅན་<span class="yigchung" id="4cdb2351350a498f8c07a84c9781835e">རྣམས་ཀྱི་སྐད་དང་མཐུན་པར་འཇུག །</span></p><p><span class="root-text" id="e79dfe48ae3940078d202c9f592ac864">ཐུགས་དང་མཛད་པ་འཕྲིན་ལས་ལྷུན་</span></p><p><span class="root-text" id="e79dfe48ae3940078d202c9f592ac8641">གྲུབ་བསམ་གྱིས་མི་ཁྱབ་བློ་ལས་རྣམ་འ</span></p><p><span class="root-text" id="e79dfe48ae3940078d202c9f592ac8642">དས་འགྲོ་བའི་དཔལ་གྱུར་ཆོས་ཀྱི་རྗེ་ལ་འདུད། །</span></p><p><span class="citation" id="a3cacfe626734376b5a45ade275de789">ཐུགས་དང་མཛད་པ་འཕྲིན་ལས་ལྷུན་</span></p><p><span class="citation" id="a3cacfe626734376b5a45ade275de7891">གྲུབ་བསམ་གྱིས་མི་ཁྱབ་བློ་ལས་རྣམ་འ</span></p><p><span class="citation" id="a3cacfe626734376b5a45ade275de7892">དས་འགྲོ་བའི་དཔལ་གྱུར་ཆོས་ཀྱི་རྗེ་ལ་འདུད། །</span></p><p><span class="sabche" id="3f56858201fc44e8808841286131a475">དུས་གསུམ་འགར་མེད་འོད་གསལ་སྒྲིབ་གསུམ་</span>རབ་སྤངས་རྣམ་དག་ཐུགས་ཀྱི་དཀྱིལ་འཁོར་བསམ་གྱིས་མི་ཁྱབ་ཡེ་ཤེས་དུས་རྣམས་ཀུ<span class="root-text" id="de30c1c91ce44217a6e0cacdcab6a404">ན་ལ་ཆགས་ཐོགས་མེད་པར་འཇུག །</span>ཡོན་ཏན་རྒྱ་མཚོ་སྐུ་གསུང་ཐུགས་དང་མཛད་པ་འཕྲིན་ལས་ལྷུན་གྲུབ་བསམ་གྱིས་མི་ཁྱབ་བློ་ལས་རྣམ་འདས་འགྲོ་བའི་དཔལ་གྱུར་ཆོས་ཀྱི་རྗེ་ལ་འདུད། །</p><p></p><p></p><p></p></body>
 </html>
