+++ tests/data/serialize/tsadra/serialized_P000801.html
@@ -1,0 +1,7 @@
+<html>
+<head>
+	<title>ཀདཕདཀདཀ</title>
+</head>
+<body>
+<p><span class="credits-page_front-title">ཀདཕདཀདཀ</span></p><p class="tibetan-regular-indented"><span class="tibetan-book-sub-title">དཕཀདཀད</span></p><p class="text-author"><span class="front-page---text-titles">དཕཀཀཀཀད</span></p><p class="credits-page_epub-edition-line"><span class="credits-regular"><img src="tests/data/serialize/tsadra/P000801/P000801.opf/assets/image/tmp6mtyug5q.png" alt="credit image not found"/></span></p><p class="tibetan-regular-indented"><span class="tibetan-book-title">༄། །འབྲི་གུང་ཚོགས་ཆོས་ཆེན་མོ།</span></p><p class="tibetan-regular-indented"><span class="tibetan-book-sub-title">༼ཚགས་ཆོས་ཀྱི་རིམ་པ་ལེགས་པར་བཤད་པ་ཡིད་བཞིན་ནོར་བུ་ཞེས་བྱ་བ་བཞུགས་སོ།།༽</span></p><p class="text-author"><span class="front-page---text-titles">ཚོགས་གཉིས་</span></p><p class="tibetan-commentary-non-indent"><span class="tibetan-citations">མཐར་ཕྱིན་མཚན་དང་དཔེ་བྱད་རིན་ཆེན་རི་བོ་ལྟར་</span>མཛེས་སྐུ་ཡི་དཀྱིལ་འཁོར་བསམ་གྱིས་མི་<span class="tibetan-citations">ཁྱབ་མཁའ་མཉམ་སེམས་ཅན་ཁམས་དང་མཉམ་པར་འཇུག །</span>ཆོས་རྣམས་གདོད་ནས་མ་སྐྱེས་སྟོང་པར་སྟོན་པའི་གསུང་གི་དཀྱིལ་འཁོར་བསམ་གྱིས་མི་ཁྱབ་ཐེ་ཚོམ་གཅོད་ཅིང་སེམས་ཅན་<span class="tibetan-commentary-small">རྣམས་ཀྱི་སྐད་དང་མཐུན་པར་འཇུག །</span></p><p class="tibetan-regular-indented"><span class="tibetan-root-text-in-verse">ཐུགས་དང་མཛད་པ་འཕྲིན་ལས་ལྷུན་</span></p><p class="tibetan-regular-indented"><span class="tibetan-root-text-in-verse-middle-line">གྲུབ་བསམ་གྱིས་མི་ཁྱབ་བློ་ལས་རྣམ་འ</span></p><p class="tibetan-regular-indented"><span class="tibetan-root-text-in-verse-last-line">དས་འགྲོ་བའི་དཔལ་གྱུར་ཆོས་ཀྱི་རྗེ་ལ་འདུད། །</span></p><p class="tibetan-regular-indented"><span class="tibetan-citations-in-verse">ཐུགས་དང་མཛད་པ་འཕྲིན་ལས་ལྷུན་</span></p><p class="tibetan-regular-indented"><span class="tibetan-citations-in-verse-middle-line">གྲུབ་བསམ་གྱིས་མི་ཁྱབ་བློ་ལས་རྣམ་འ</span></p><p class="tibetan-regular-indented"><span class="tibetan-citations-in-verse-last-line">དས་འགྲོ་བའི་དཔལ་གྱུར་ཆོས་ཀྱི་རྗེ་ལ་འདུད། །</span></p><p class="tibetan-commentary-non-indent"><span class="tibetan-sabche">དུས་གསུམ་འགར་མེད་འོད་གསལ་སྒྲིབ་གསུམ་</span>རབ་སྤངས་རྣམ་དག་ཐུགས་ཀྱི་དཀྱིལ་འཁོར་བསམ་གྱིས་མི་ཁྱབ་ཡེ་ཤེས་དུས་རྣམས་ཀུ<span class="tibetan-root-text">ན་ལ་ཆགས་ཐོགས་མེད་པར་འཇུག །</span>ཡོན་ཏན་རྒྱ་མཚོ་སྐུ་གསུང་ཐུགས་དང་མཛད་པ་འཕྲིན་ལས་ལྷུན་གྲུབ་བསམ་གྱིས་མི་ཁྱབ་བློ་ལས་རྣམ་འདས་འགྲོ་བའི་དཔལ་གྱུར་ཆོས་ཀྱི་རྗེ་ལ་འདུད། །</p><p class="tibetan-regular-indented"></p><p class="tibetan-regular-indented"></p></body>
+</html>
